[Core] Moving the MortarUtilities to cpp
This reduces linking and compilation time when something is changed
