+++ kratos/utilities/mortar_utilities.h
@@ -20,8 +20,6 @@
 // External includes
 
 // Project includes
-#include "utilities/math_utils.h"
-#include "utilities/variable_utils.h"
 #include "includes/enums.h"
 #include "includes/model_part.h"
 #include "utilities/openmp_utils.h"
@@ -48,7 +46,7 @@ namespace Kratos
 ///@{
 
 /**
- * @class MortarUtilities
+ * @namespace MortarUtilities
  * @ingroup KratosCore
  * @brief This is a class that provides auxiliar utilities for the mortar integration
  * @details This is a class that provides auxiliar utilities for the mortar integration. Many methods
@@ -57,14 +55,10 @@ namespace Kratos
  * @author Vicente Mataix Ferrandiz
  * Contact: vmataix@cimne.upc.edu
  */
-class MortarUtilities
+namespace MortarUtilities
 {
-public:
     ///@name Type Definitions
     ///@{
-
-    /// Pointer definition of MortarUtilities
-    KRATOS_CLASS_POINTER_DEFINITION( MortarUtilities );
 
     // Some geometrical definitions
     typedef Node<3>                                              NodeType;
@@ -74,13 +68,6 @@ public:
     /// Definition of geometries
     typedef Geometry<NodeType>                               GeometryType;
     typedef Geometry<PointType>                         GeometryPointType;
-
-    /// The integration method type
-    typedef GeometryData::IntegrationMethod             IntegrationMethod;
-
-    /// The containers of the components of the model parts
-    typedef ModelPart::NodesContainerType                  NodesArrayType;
-    typedef ModelPart::ConditionsContainerType        ConditionsArrayType;
 
     /// Index type definition
     typedef std::size_t                                         IndexType;
@@ -92,33 +79,7 @@ public:
     typedef std::unordered_map<IndexType, IndexType>               IntMap;
 
     ///@}
-    ///@name Life Cycle
-    ///@{
-
-    /// Default constructor.
-    MortarUtilities(){}
-
-    /// Destructor.
-    virtual ~MortarUtilities(){}
-
-    ///@}
-    ///@name Access
-    ///@{
-
-    ///@}
-    ///@name Inquiry
-    ///@{
-
-    ///@}
-    ///@name Input and output
-    ///@{
-
-    ///@}
-    ///@name Friends
-    ///@{
-
-    ///@}
-    ///@name Operations
+    ///@name  Functions
     ///@{
 
     /**
@@ -127,26 +88,17 @@ public:
      * @param Tolerance The threshold length
      * @return True if the line is too short, false otherwise
      */
-    static inline bool LengthCheck(
+    bool KRATOS_API(KRATOS_CORE) LengthCheck(
         const GeometryPointType& rGeometryLine,
         const double Tolerance = 1.0e-6
-        ) {
-        const double lx = rGeometryLine[0].X() - rGeometryLine[1].X();
-        const double ly = rGeometryLine[0].Y() - rGeometryLine[1].Y();
-
-        const double length = std::sqrt(lx * lx + ly * ly);
-
-        return (length < Tolerance) ? true : false;
-    }
+        );
 
     /**
      * @brief This functions checks if the semiperimeter is smaller than any of the sides of the triangle
      * @param rGeometryTriangle The triangle to be checked
      * @return True if the triangle is in bad shape, false otherwise
      */
-    static inline bool HeronCheck(const GeometryPointType& rGeometryTriangle) {
-        return HeronCheck(rGeometryTriangle[0], rGeometryTriangle[1], rGeometryTriangle[2]);
-    }
+    bool KRATOS_API(KRATOS_CORE) HeronCheck(const GeometryPointType& rGeometryTriangle);
 
     /**
      * @brief This functions checks if the semiperimeter is smaller than any of the sides of the triangle
@@ -155,30 +107,11 @@ public:
      * @param rPointOrig3 The triangle third point
      * @return True if the triangle is in bad shape, false otherwise
      */
-    static inline bool HeronCheck(
+    bool KRATOS_API(KRATOS_CORE) HeronCheck(
         const PointType& rPointOrig1,
         const PointType& rPointOrig2,
         const PointType& rPointOrig3
-        )
-    {
-        const double a = MathUtils<double>::Norm3(rPointOrig1.Coordinates()-rPointOrig2.Coordinates());
-        const double b = MathUtils<double>::Norm3(rPointOrig2.Coordinates()-rPointOrig3.Coordinates());
-        const double c = MathUtils<double>::Norm3(rPointOrig3.Coordinates()-rPointOrig1.Coordinates());
-
-        const double s = 0.5 * (a + b + c);
-        const double A2 = s * (s - a) * (s - b) * (s - c);
-
-        const bool Check = A2 <= 0.0 ? true : false;  // We consider as bad shaped the ones with no area or negative A2 (semiperimeter smaller than any side)
-
-//         // Debug
-//         KRATOS_INFO("Check") << Check << " A2: " << A2 << std::endl;
-//         if (Check == true) {
-//             KRATOS_WARNING("Bad shape") << "Warning:: The triangle is in bad shape" << std::endl;
-//             KRATOS_INFO("Mathematica triangle") << "Graphics3D[{EdgeForm[Thick],Triangle[{{" << rPointOrig1.X() << "," << rPointOrig1.Y() << "," << rPointOrig1.Z()  << "},{" << rPointOrig2.X() << "," << rPointOrig2.Y() << "," << rPointOrig2.Z()  << "},{" << rPointOrig3.X() << "," << rPointOrig3.Y() << "," << rPointOrig3.Z()  << "}}]}]" << std::endl;
-//         }
-
-        return Check;
-    }
+        );
 
     /**
      * @brief This function rotates to align the projected points to a parallel plane to XY
@@ -188,34 +121,13 @@ public:
      * @param rSlaveTangentEta The second tangent vector of the slave condition
      * @param Inversed If we rotate to the XY or we recover from XY
      */
-    static inline void RotatePoint(
+    void KRATOS_API(KRATOS_CORE) RotatePoint(
         PointType& rPointToRotate,
         const PointType& rPointReferenceRotation,
         const array_1d<double, 3>& rSlaveTangentXi,
         const array_1d<double, 3>& rSlaveTangentEta,
         const bool Inversed
-        )
-    {
-        // We move to the (0,0,0)
-        PointType aux_point_to_rotate;
-        aux_point_to_rotate.Coordinates() = rPointToRotate.Coordinates() - rPointReferenceRotation.Coordinates();
-
-        BoundedMatrix<double, 3, 3> rotation_matrix = ZeroMatrix(3, 3);
-
-        if (Inversed == false) {
-            for (IndexType i = 0; i < 3; ++i) {
-                rotation_matrix(0, i) = rSlaveTangentXi[i];
-                rotation_matrix(1, i) = rSlaveTangentEta[i];
-            }
-        } else {
-            for (IndexType i = 0; i < 3; ++i) {
-                rotation_matrix(i, 0) = rSlaveTangentXi[i];
-                rotation_matrix(i, 1) = rSlaveTangentEta[i];
-            }
-        }
-
-        rPointToRotate.Coordinates() = prod(rotation_matrix, aux_point_to_rotate) + rPointReferenceRotation.Coordinates();
-    }
+        );
 
     /**
      * @brief This function calculates the r_normal in a specific GP with a given shape function
@@ -223,22 +135,10 @@ public:
      * @param rGeometry The geometry of condition of interest
      * @return The r_normal in the GP
      */
-    static inline array_1d<double,3> GaussPointUnitNormal(
+    array_1d<double,3> KRATOS_API(KRATOS_CORE) GaussPointUnitNormal(
         const Vector& rN,
         const GeometryType& rGeometry
-        ) {
-        array_1d<double,3> r_normal = ZeroVector(3);
-        for( IndexType i_node = 0; i_node < rGeometry.PointsNumber(); ++i_node )
-            r_normal += rN[i_node] * rGeometry[i_node].FastGetSolutionStepValue(NORMAL);
-
-        const double this_norm = norm_2(r_normal);
-
-        KRATOS_DEBUG_ERROR_IF(this_norm < std::numeric_limits<double>::epsilon()) << "Zero norm r_normal vector. Norm:" << this_norm << std::endl;
-
-        r_normal /= this_norm;
-
-        return r_normal;
-    }
+        );
 
     /**
      * @brief This function gives you the indexes needed to order a vector
@@ -246,7 +146,7 @@ public:
      * @return idx The vector of indexes
      */
     template <typename TType>
-    static std::vector<std::size_t> SortIndexes(const std::vector<TType> &rThisVector) {
+    std::vector<std::size_t> SortIndexes(const std::vector<TType> &rThisVector) {
         // Initialize original index locations
         std::vector<std::size_t> idx(rThisVector.size());
         iota(idx.begin(), idx.end(), 0);
@@ -262,67 +162,14 @@ public:
      * @brief It computes the mean of the r_normal in the condition in all the nodes
      * @param rModelPart The model part to compute
      */
-    static inline void ComputeNodesMeanNormalModelPart(ModelPart& rModelPart) {
-        // Check NORMAL is available
-        KRATOS_ERROR_IF_NOT(rModelPart.HasNodalSolutionStepVariable(NORMAL)) << "NORMAL is not available on the solution step data variable database" << std::endl;
-        
-        // We iterate over nodes
-        NodesArrayType& r_nodes_array = rModelPart.Nodes();
-        const auto it_node_begin = r_nodes_array.begin();
-        const int num_nodes = static_cast<int>(r_nodes_array.size());
-
-        // Auxiliar zero array
-        const array_1d<double, 3> zero_array = ZeroVector(3);
-
-        // Reset NORMAL
-        VariableUtils().SetVectorVar(NORMAL, zero_array, r_nodes_array);
-
-        // Sum all the nodes normals
-        ConditionsArrayType& r_conditions_array = rModelPart.Conditions();
-        const auto it_cond_begin = r_conditions_array.begin();
-
-        // Declare auxiliar coordinates
-        CoordinatesArrayType aux_coords;
-
-        #pragma omp parallel for firstprivate(aux_coords)
-        for(int i = 0; i < static_cast<int>(r_conditions_array.size()); ++i) {
-            auto it_cond = it_cond_begin + i;
-            const GeometryType& r_geometry = it_cond->GetGeometry();
-
-            // Set condition normal
-            r_geometry.PointLocalCoordinates(aux_coords, r_geometry.Center());
-            it_cond->SetValue(NORMAL, r_geometry.UnitNormal(aux_coords));
-        }
-
-        // Adding the normal contribution of each node
-        for(Condition& r_cond : r_conditions_array) {
-            GeometryType& r_geometry = r_cond.GetGeometry();
-
-            // Iterate over nodes
-            for (NodeType& r_node : r_geometry) {
-                r_geometry.PointLocalCoordinates(aux_coords, r_node.Coordinates());
-                noalias(r_node.FastGetSolutionStepValue(NORMAL)) += r_geometry.UnitNormal(aux_coords);
-            }
-        }
-
-        #pragma omp parallel for
-        for(int i = 0; i < num_nodes; ++i) {
-            auto it_node = it_node_begin + i;
-
-            array_1d<double, 3>& r_normal = it_node->FastGetSolutionStepValue(NORMAL);
-            const double norm_normal = norm_2(r_normal);
-
-            if (norm_normal > std::numeric_limits<double>::epsilon()) r_normal /= norm_normal;
-            else KRATOS_ERROR_IF(it_node->Is(INTERFACE)) << "ERROR:: ZERO NORM NORMAL IN NODE: " << it_node->Id() << std::endl;
-        }
-    }
+    void KRATOS_API(KRATOS_CORE) ComputeNodesMeanNormalModelPart(ModelPart& rModelPart);
 
     /**
      * @brief It inverts the order of the nodes in the conditions of a model part in order to invert the r_normal
      * @param rContainer reference to the objective container
      */
     template<class TContainerType>
-    static inline void InvertNormal(TContainerType& rContainer) {
+    void InvertNormal(TContainerType& rContainer) {
         const auto it_cont_begin = rContainer.begin();
         #pragma omp parallel for
         for(int i = 0; i < static_cast<int>(rContainer.size()); ++i) {
@@ -343,7 +190,7 @@ public:
      */
 
     template< SizeType TDim, SizeType TNumNodes>
-    static inline BoundedMatrix<double, TNumNodes, TDim> GetCoordinates(
+    BoundedMatrix<double, TNumNodes, TDim> GetCoordinates(
         const GeometryType& rGeometry,
         const bool Current = true,
         const IndexType Step = 0
@@ -375,7 +222,7 @@ public:
      * @return tangent_matrix The matrix containing the tangent vectors of the LM
      */
     template< SizeType TNumNodes, SizeType TDim>
-    static inline BoundedMatrix<double, TNumNodes, TDim> ComputeTangentMatrix(const GeometryType& rGeometry) {
+    BoundedMatrix<double, TNumNodes, TDim> ComputeTangentMatrix(const GeometryType& rGeometry) {
         /* DEFINITIONS */
         // Zero tolerance
         const double zero_tolerance = std::numeric_limits<double>::epsilon();
@@ -412,7 +259,7 @@ public:
      * @return var_vector The vector containing the variables of the geometry
      */
     template< SizeType TNumNodes, class TVarType = Variable<double>>
-    static inline array_1d<double, TNumNodes> GetVariableVector(
+    array_1d<double, TNumNodes> GetVariableVector(
         const GeometryType& rGeometry,
         const TVarType& rVariable,
         const IndexType Step
@@ -434,7 +281,7 @@ public:
      * @return var_vector The vector containing the variables of the geometry
      */
     template< SizeType TNumNodes, class TVarType = Variable<double> >
-    static inline BoundedMatrix<double, TNumNodes, 1> GetVariableVectorMatrix(
+    BoundedMatrix<double, TNumNodes, 1> GetVariableVectorMatrix(
         const GeometryType& rGeometry,
         const TVarType& rVariable,
         const unsigned int Step
@@ -455,7 +302,7 @@ public:
      * @return var_vector The vector containing the variables of the geometry
      */
     template< SizeType TNumNodes, class TVarType = Variable<double> >
-    static inline array_1d<double, TNumNodes> GetVariableVector(
+    array_1d<double, TNumNodes> GetVariableVector(
         const GeometryType& rGeometry,
         const TVarType& rVariable
         ) {
@@ -475,7 +322,7 @@ public:
      * @return var_vector The vector containing the variables of the geometry
      */
     template< SizeType TNumNodes, class TVarType = Variable<double> >
-    static inline BoundedMatrix<double, TNumNodes, 1> GetVariableVectorMatrix(
+    BoundedMatrix<double, TNumNodes, 1> GetVariableVectorMatrix(
         const GeometryType& rGeometry,
         const TVarType& rVariable
         ) {
@@ -496,7 +343,7 @@ public:
      * @return var_matrix The matrix containing the variables of the geometry
      */
     template< SizeType TDim, SizeType TNumNodes>
-    static inline BoundedMatrix<double, TNumNodes, TDim> GetVariableMatrix(
+    BoundedMatrix<double, TNumNodes, TDim> GetVariableMatrix(
         const GeometryType& rGeometry,
         const Variable<array_1d<double,3> >& rVariable,
         const unsigned int Step
@@ -520,7 +367,7 @@ public:
      * @return var_matrix The matrix containing the variables of the geometry
      */
     template< SizeType TDim, SizeType TNumNodes>
-    static inline BoundedMatrix<double, TNumNodes, TDim> GetVariableMatrix(
+    BoundedMatrix<double, TNumNodes, TDim> GetVariableMatrix(
         const GeometryType& rGeometry,
         const Variable<array_1d<double,3> >& rVariable
         ) {
@@ -542,7 +389,7 @@ public:
      * @return AbsMatrix The matrix containing the absolute value of another matrix
      */
     template< SizeType TDim, SizeType TNumNodes>
-    static inline BoundedMatrix<double, TNumNodes, TDim> GetAbsMatrix(const BoundedMatrix<double, TNumNodes, TDim>& rInputMatrix) {
+    BoundedMatrix<double, TNumNodes, TDim> GetAbsMatrix(const BoundedMatrix<double, TNumNodes, TDim>& rInputMatrix) {
         /* DEFINITIONS */
         BoundedMatrix<double, TNumNodes, TDim> AbsMatrix;
 
@@ -558,7 +405,7 @@ public:
      * @brief This method gives the size to be computed
      */
     template< SizeType TDim, class TVarType>
-    static inline unsigned int SizeToCompute()
+    unsigned int SizeToCompute()
     {
        if (typeid(TVarType) == typeid(Variable<array_1d<double, 3>>))
            return TDim;
@@ -572,7 +419,7 @@ public:
      * @param rThisVariable The variable to set
      */
     template< class TVarType, HistoricalValues THist>
-    static inline void ResetValue(
+    void KRATOS_API(KRATOS_CORE) ResetValue(
         ModelPart& rThisModelPart,
         TVarType& rThisVariable
         );
@@ -582,19 +429,23 @@ public:
      * @param rThisModelPart The model part to update
      */
     template< class TVarType>
-    static inline void ResetAuxiliarValue(ModelPart& rThisModelPart);
+    void KRATOS_API(KRATOS_CORE) ResetAuxiliarValue(ModelPart& rThisModelPart);
 
     /**
      * @brief This method returns the auxiliar variable
+     * @return The auxiliar variable
      */
     template< class TVarType>
-    static inline TVarType GetAuxiliarVariable();
+    TVarType KRATOS_API(KRATOS_CORE) GetAuxiliarVariable();
 
     /**
      * @brief This method returns the auxiliar variable
+     * @param pThisNode Pointer to the node of interest
+     * @param iSize The Index of the component
+     * @return The value of the auxiliar variable
      */
     template< class TVarType>
-    static inline double GetAuxiliarValue(
+    double KRATOS_API(KRATOS_CORE) GetAuxiliarValue(
         NodeType::Pointer pThisNode,
         unsigned int iSize
         );
@@ -606,7 +457,7 @@ public:
      * @param rThisValue The matrix to be updated
      */
     template< class TVarType, HistoricalValues THist>
-    static inline void MatrixValue(
+    void KRATOS_API(KRATOS_CORE) MatrixValue(
         GeometryType& rThisGeometry,
         TVarType& rThisVariable,
         Matrix& rThisValue
@@ -620,7 +471,7 @@ public:
      * @param rThisValue The matrix to be updated
      */
     template< class TVarType, HistoricalValues THist>
-    static inline void AddValue(
+    void KRATOS_API(KRATOS_CORE) AddValue(
         GeometryType& rThisGeometry,
         TVarType& rThisVariable,
         const Matrix& rThisValue
@@ -632,7 +483,7 @@ public:
      * @param rThisVariable The variable to set
      */
     template< class TVarType, HistoricalValues THist>
-    static inline void AddAreaWeightedNodalValue(
+    void KRATOS_API(KRATOS_CORE) AddAreaWeightedNodalValue(
         NodeType::Pointer pThisNode,
         TVarType& rThisVariable,
         const double RefArea = 1.0,
@@ -648,366 +499,13 @@ public:
      * @param ConectivityDatabase The database that will be used to assemble the system
      */
     template< class TVarType, HistoricalValues THist>
-    static inline void UpdateDatabase(
+    void KRATOS_API(KRATOS_CORE) UpdateDatabase(
         ModelPart& rThisModelPart,
         TVarType& rThisVariable,
         Vector& Dx,
         unsigned int Index,
         IntMap& ConectivityDatabase
         );
-
-private:
-};// class MortarUtilities
-
-///@name Explicit Specializations
-///@{
-
-template<>
-inline void MortarUtilities::ResetValue<Variable<double>, Historical>(
-        ModelPart& rThisModelPart,
-        Variable<double>& rThisVariable
-        )
-{
-    NodesArrayType& r_nodes_array = rThisModelPart.Nodes();
-    VariableUtils().SetScalarVar(rThisVariable, 0.0, r_nodes_array);
-}
-
-template<>
-inline void MortarUtilities::ResetValue<Variable<array_1d<double, 3>>, Historical>(
-        ModelPart& rThisModelPart,
-        Variable<array_1d<double, 3>>& rThisVariable
-        ) {
-    NodesArrayType& r_nodes_array = rThisModelPart.Nodes();
-    VariableUtils().SetVectorVar(rThisVariable, ZeroVector(3), r_nodes_array);
-}
-
-template<>
-inline void MortarUtilities::ResetValue<Variable<double>, NonHistorical>(
-        ModelPart& rThisModelPart,
-        Variable<double>& rThisVariable
-        ) {
-    NodesArrayType& r_nodes_array = rThisModelPart.Nodes();
-    VariableUtils().SetNonHistoricalVariable(rThisVariable, 0.0, r_nodes_array);
-}
-
-template<>
-inline void MortarUtilities::ResetValue<Variable<array_1d<double, 3>>, NonHistorical>(
-        ModelPart& rThisModelPart,
-        Variable<array_1d<double, 3>>& rThisVariable
-        ) {
-    const array_1d<double, 3> zero_array = ZeroVector(3);
-    NodesArrayType& r_nodes_array = rThisModelPart.Nodes();
-    VariableUtils().SetNonHistoricalVariable(rThisVariable, zero_array, r_nodes_array);
-}
-
-template<>
-inline void MortarUtilities::ResetAuxiliarValue<Variable<double>>(ModelPart& rThisModelPart) {
-    NodesArrayType& r_nodes_array = rThisModelPart.Nodes();
-    VariableUtils().SetNonHistoricalVariable(NODAL_MAUX, 0.0, r_nodes_array);
-}
-
-template<>
-inline void MortarUtilities::ResetAuxiliarValue<Variable<array_1d<double, 3>>>(ModelPart& rThisModelPart) {
-    const array_1d<double, 3> zero_array = ZeroVector(3);
-    NodesArrayType& r_nodes_array = rThisModelPart.Nodes();
-    VariableUtils().SetNonHistoricalVariable(NODAL_VAUX, zero_array, r_nodes_array);
-}
-
-template< >
-inline Variable<double> MortarUtilities::GetAuxiliarVariable<Variable<double>>() {
-    return NODAL_MAUX;
-}
-
-template< >
-inline Variable<array_1d<double, 3>> MortarUtilities::GetAuxiliarVariable<Variable<array_1d<double, 3>>>() {
-    return NODAL_VAUX;
-}
-
-template< >
-inline double MortarUtilities::GetAuxiliarValue<Variable<double>>(
-    Node<3>::Pointer pThisNode,
-    unsigned int iSize
-    ) {
-    return pThisNode->GetValue(NODAL_MAUX);
-}
-
-template< >
-inline double MortarUtilities::GetAuxiliarValue<Variable<array_1d<double, 3>>>(
-    Node<3>::Pointer pThisNode,
-    unsigned int iSize
-    ) {
-    switch ( iSize ) {
-        case 0:
-            return pThisNode->GetValue(NODAL_VAUX_X);
-        case 1:
-            return pThisNode->GetValue(NODAL_VAUX_Y);
-        case 2:
-            return pThisNode->GetValue(NODAL_VAUX_Z);
-        default:
-            return 0.0;
-    }
-
-    return 0.0;
-}
-
-template<>
-inline void MortarUtilities::MatrixValue<Variable<double>, Historical>(
-        GeometryType& rThisGeometry,
-        Variable<double>& rThisVariable,
-        Matrix& rThisValue
-        ) {
-    if (rThisValue.size1() != rThisGeometry.size() || rThisValue.size2() != 1)
-        rThisValue.resize(rThisGeometry.size(), 1, false);
-
-    for (IndexType i_node = 0; i_node < rThisGeometry.size(); ++i_node)
-        rThisValue(i_node, 0) = rThisGeometry[i_node].FastGetSolutionStepValue(rThisVariable);
-}
-
-template<>
-inline void MortarUtilities::MatrixValue<Variable<array_1d<double, 3>>, Historical>(
-        GeometryType& rThisGeometry,
-        Variable<array_1d<double, 3>>& rThisVariable,
-        Matrix& rThisValue
-        ) {
-    const std::size_t num_nodes = rThisGeometry.size();
-    const std::size_t dimension = rThisGeometry.WorkingSpaceDimension();
-    if (rThisValue.size1() != num_nodes || rThisValue.size2() != dimension)
-        rThisValue.resize(num_nodes, dimension, false);
-
-    for (IndexType i_node = 0; i_node < num_nodes; ++i_node) {
-        const auto& rvalue = rThisGeometry[i_node].FastGetSolutionStepValue(rThisVariable);
-        for (IndexType i_dim = 0; i_dim < dimension; ++i_dim)
-            rThisValue(i_node, i_dim) = rvalue[i_dim];
-    }
-}
-template<>
-inline void MortarUtilities::MatrixValue<Variable<double>, NonHistorical>(
-        GeometryType& rThisGeometry,
-        Variable<double>& rThisVariable,
-        Matrix& rThisValue
-        ) {
-    if (rThisValue.size1() != rThisGeometry.size() || rThisValue.size2() != 1)
-        rThisValue.resize(rThisGeometry.size(), 1, false);
-
-    for (IndexType i_node = 0; i_node < rThisGeometry.size(); ++i_node)
-        rThisValue(i_node, 0) = rThisGeometry[i_node].GetValue(rThisVariable);
-}
-
-template<>
-inline void MortarUtilities::MatrixValue<Variable<array_1d<double, 3>>, NonHistorical>(
-        GeometryType& rThisGeometry,
-        Variable<array_1d<double, 3>>& rThisVariable,
-        Matrix& rThisValue
-        ) {
-    const std::size_t num_nodes = rThisGeometry.size();
-    const std::size_t dimension = rThisGeometry.WorkingSpaceDimension();
-    if (rThisValue.size1() != num_nodes || rThisValue.size2() != dimension)
-        rThisValue.resize(num_nodes, dimension, false);
-
-    for (IndexType i_node = 0; i_node < num_nodes; ++i_node) {
-        const auto& rvalue = rThisGeometry[i_node].GetValue(rThisVariable);
-        for (IndexType i_dim = 0; i_dim < dimension; ++i_dim)
-            rThisValue(i_node, i_dim) = rvalue[i_dim];
-    }
-}
-
-template<>
-inline void MortarUtilities::AddValue<Variable<double>, Historical>(
-        GeometryType& rThisGeometry,
-        Variable<double>& rThisVariable,
-        const Matrix& rThisValue
-        ) {
-    for (IndexType i_node = 0; i_node < rThisGeometry.size(); ++i_node) {
-        double& aux_value = rThisGeometry[i_node].FastGetSolutionStepValue(rThisVariable);
-        #pragma omp atomic
-        aux_value += rThisValue(i_node, 0);
-    }
-}
-
-template<>
-inline void MortarUtilities::AddValue<Variable<array_1d<double, 3>>, Historical>(
-        GeometryType& rThisGeometry,
-        Variable<array_1d<double, 3>>& rThisVariable,
-        const Matrix& rThisValue
-        ) {
-    for (IndexType i_node = 0; i_node < rThisGeometry.size(); ++i_node) {
-        auto& aux_vector = rThisGeometry[i_node].FastGetSolutionStepValue(rThisVariable);
-        for (unsigned int i_dim = 0; i_dim < rThisGeometry.WorkingSpaceDimension(); ++i_dim) {
-            double& aux_value = aux_vector[i_dim];
-            #pragma omp atomic
-            aux_value += rThisValue(i_node, i_dim);
-        }
-    }
-}
-template<>
-inline void MortarUtilities::AddValue<Variable<double>, NonHistorical>(
-        GeometryType& rThisGeometry,
-        Variable<double>& rThisVariable,
-        const Matrix& rThisValue
-        ) {
-    for (IndexType i_node = 0; i_node < rThisGeometry.size(); ++i_node) {
-        double& aux_value = rThisGeometry[i_node].GetValue(rThisVariable);
-        #pragma omp atomic
-        aux_value += rThisValue(i_node, 0);
-    }
-}
-
-template<>
-inline void MortarUtilities::AddValue<Variable<array_1d<double, 3>>, NonHistorical>(
-        GeometryType& rThisGeometry,
-        Variable<array_1d<double, 3>>& rThisVariable,
-        const Matrix& rThisValue
-        ) {
-    for (IndexType i_node = 0; i_node < rThisGeometry.size(); ++i_node) {
-        auto& aux_vector = rThisGeometry[i_node].GetValue(rThisVariable);
-        for (unsigned int i_dim = 0; i_dim < rThisGeometry.WorkingSpaceDimension(); ++i_dim) {
-            double& aux_value = aux_vector[i_dim];
-            #pragma omp atomic
-            aux_value += rThisValue(i_node, i_dim);
-        }
-    }
-}
-
-template<>
-inline void MortarUtilities::AddAreaWeightedNodalValue<Variable<double>, Historical>(
-        Node<3>::Pointer pThisNode,
-        Variable<double>& rThisVariable,
-        const double RefArea,
-        const double Tolerance
-        ) {
-    double area_coeff = pThisNode->GetValue(NODAL_AREA);
-    const bool null_area = (std::abs(area_coeff) < RefArea * Tolerance);
-#ifdef KRATOS_DEBUG
-    if (null_area) KRATOS_WARNING("WARNING:: NODE OF NULL AREA.") << " ID: " << pThisNode->Id() << std::endl;
-#endif
-    area_coeff = null_area ? 0.0 : 1.0/area_coeff;
-    double& aux_value = pThisNode->FastGetSolutionStepValue(rThisVariable);
-    #pragma omp atomic
-    aux_value += area_coeff * pThisNode->GetValue(NODAL_MAUX);
-}
-
-template<>
-inline void MortarUtilities::AddAreaWeightedNodalValue<Variable<array_1d<double, 3>>, Historical>(
-        Node<3>::Pointer pThisNode,
-        Variable<array_1d<double, 3>>& rThisVariable,
-        const double RefArea,
-        const double Tolerance
-        ) {
-    double area_coeff = pThisNode->GetValue(NODAL_AREA);
-    const bool null_area = (std::abs(area_coeff) < RefArea * Tolerance);
-#ifdef KRATOS_DEBUG
-    if (null_area) KRATOS_WARNING("WARNING:: NODE OF NULL AREA.") << " ID: " << pThisNode->Id() << std::endl;
-#endif
-    area_coeff = null_area ? 0.0 : 1.0/area_coeff;
-    auto& aux_vector = pThisNode->FastGetSolutionStepValue(rThisVariable);
-    const auto& nodal_vaux = pThisNode->GetValue(NODAL_VAUX);
-    for (IndexType i = 0; i < 3; ++i) {
-        double& aux_value = aux_vector[i];
-        #pragma omp atomic
-        aux_value += area_coeff * nodal_vaux[i];
-    }
-}
-
-template<>
-inline void MortarUtilities::AddAreaWeightedNodalValue<Variable<double>, NonHistorical>(
-        Node<3>::Pointer pThisNode,
-        Variable<double>& rThisVariable,
-        const double RefArea,
-        const double Tolerance
-        ) {
-    double area_coeff = pThisNode->GetValue(NODAL_AREA);
-    const bool null_area = (std::abs(area_coeff) < RefArea * Tolerance);
-#ifdef KRATOS_DEBUG
-    if (null_area) KRATOS_WARNING("WARNING:: NODE OF NULL AREA.") << " ID: " << pThisNode->Id() << std::endl;
-#endif
-    area_coeff = null_area ? 0.0 : 1.0/area_coeff;
-    double& aux_value = pThisNode->GetValue(rThisVariable);
-    #pragma omp atomic
-    aux_value += area_coeff * pThisNode->GetValue(NODAL_MAUX);
-}
-
-template<>
-inline void MortarUtilities::AddAreaWeightedNodalValue<Variable<array_1d<double, 3>>, NonHistorical>(
-        Node<3>::Pointer pThisNode,
-        Variable<array_1d<double, 3>>& rThisVariable,
-        const double RefArea,
-        const double Tolerance
-        ) {
-    double area_coeff = pThisNode->GetValue(NODAL_AREA);
-    const bool null_area = (std::abs(area_coeff) < RefArea * Tolerance);
-#ifdef KRATOS_DEBUG
-    if (null_area) KRATOS_WARNING("WARNING:: NODE OF NULL AREA.") << " ID: " << pThisNode->Id() << std::endl;
-#endif
-    area_coeff = null_area ? 0.0 : 1.0/area_coeff;
-    auto& aux_vector = pThisNode->GetValue(rThisVariable);
-    const auto& nodal_vaux = pThisNode->GetValue(NODAL_VAUX);
-    for (IndexType i = 0; i < 3; ++i) {
-        double& aux_value = aux_vector[i];
-        #pragma omp atomic
-        aux_value += area_coeff * nodal_vaux[i];
-    }
-}
-
-template<>
-inline void MortarUtilities::UpdateDatabase<Variable<double>, Historical>(
-        ModelPart& rThisModelPart,
-        Variable<double>& rThisVariable,
-        Vector& Dx,
-        unsigned int Index,
-        IntMap& ConectivityDatabase
-        ) {
-    #pragma omp parallel for
-    for (int i = 0; i < static_cast<int>(Dx.size()); ++i) {
-        auto p_node = rThisModelPart.pGetNode(ConectivityDatabase[i]);
-        p_node->FastGetSolutionStepValue(rThisVariable) += Dx[i];
-    }
-}
-
-template<>
-inline void MortarUtilities::UpdateDatabase<Variable<array_1d<double, 3>>, Historical>(
-        ModelPart& rThisModelPart,
-        Variable<array_1d<double, 3>>& rThisVariable,
-        Vector& Dx,
-        unsigned int Index,
-        IntMap& ConectivityDatabase
-        ) {
-    #pragma omp parallel for
-    for (int i = 0; i < static_cast<int>(Dx.size()); ++i) {
-        auto p_node = rThisModelPart.pGetNode(ConectivityDatabase[i]);
-        auto& value = p_node->FastGetSolutionStepValue(rThisVariable);
-        value[Index] += Dx[i];
-    }
-}
-template<>
-inline void MortarUtilities::UpdateDatabase<Variable<double>, NonHistorical>(
-        ModelPart& rThisModelPart,
-        Variable<double>& rThisVariable,
-        Vector& Dx,
-        unsigned int Index,
-        IntMap& ConectivityDatabase
-        ) {
-    #pragma omp parallel for
-    for (int i = 0; i < static_cast<int>(Dx.size()); ++i) {
-        auto p_node = rThisModelPart.pGetNode(ConectivityDatabase[i]);
-        p_node->GetValue(rThisVariable) += Dx[i];
-    }
-}
-
-template<>
-inline void MortarUtilities::UpdateDatabase<Variable<array_1d<double, 3>>, NonHistorical>(
-        ModelPart& rThisModelPart,
-        Variable<array_1d<double, 3>>& rThisVariable,
-        Vector& Dx,
-        unsigned int Index,
-        IntMap& ConectivityDatabase
-        ) {
-    #pragma omp parallel for
-    for (int i = 0; i < static_cast<int>(Dx.size()); ++i) {
-        auto p_node = rThisModelPart.pGetNode(ConectivityDatabase[i]);
-        auto& value = p_node->GetValue(rThisVariable);
-        value[Index] += Dx[i];
-    }
-}
-
-}
+};// namespace MortarUtilities
+} // namespace Kratos
 #endif /* KRATOS_MORTAR_UTILITIES defined */
